refactor(tauri) use webview_init to inject the init script
